% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisshu_cross.R
\name{jisshu_cross}
\alias{jisshu_cross}
\title{2 way cross table for chosa jisshu}
\usage{
jisshu_cross(.data, .x, .y)
}
\arguments{
\item{.data}{a \code{tibble} or \code{data.frame}}

\item{.x}{a column name in \code{.data}}

\item{.y}{a column name in \code{.data}}
}
\value{
an object of class \code{wbWorkbook} exported from \code{openxlsx2} package. The object can be easily saved as an excel file. See examples for details.
}
\description{
調査実習用の2次元クロス表を作成する関数です。
}
\details{
データと変数2つを指定するとクロス表が作成されます。

クラメールのVとカイ2乗検定のp値も算出されます。2×2のクロス表の場合はイェーツの連続修正が入ります。

エクセルへの保存はexampleを参照してください。
}
\examples{
# Create a sample data
data <-
  data.frame(
    x = c(rep('A', 50), rep('B', 50)),
    y = c(rep(1:5, 20))
  )

# Create a cross table
result <- data |> jisshu_cross(x, y)

# Save the cross table as an excel file
\dontrun{
result$save('hoge.xlsx')
}

}
