% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisshu_reg.R
\name{jisshu_reg}
\alias{jisshu_reg}
\alias{jisshu_reg.lm}
\title{regression table for tyosa jisshu}
\usage{
jisshu_reg(object)

\method{jisshu_reg}{lm}(object)
}
\arguments{
\item{object}{an object for which a summary is desired. For example, a model object, such as those returned by \code{lm} and \code{glm}.}
}
\description{
調査実習用の回帰分析結果を作成する関数です。
}
\details{
\code{lm}や\code{glm}の結果から、執筆要項に沿ったエクセルの表が作成できます。

エクセルへの保存はexampleを参照してください。
}
\examples{
# Create a sample data
data <- tibble::tibble(
  x1 = rnorm(100, mean = 0, sd = 1),
  x2 = rnorm(100, mean = 0, sd = 1),
  y = 3 + 0.3*x1 + 0.5*x2 + rnorm(100, mean = 0, sd = 0.1)
)

# Create a regression table
model <- lm(y ~ x1 + x2, data = data)
result <- jisshu_reg(model)

# Save the regression table as an excel file
\dontrun{
result$save('hoge.xlsx')
}

}
