% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisshu_simple_tabulation.R
\name{jisshu_simple_tabulation}
\alias{jisshu_simple_tabulation}
\title{Simple Tabulation for Chosa Jisshu}
\usage{
jisshu_simple_tabulation(
  data,
  path,
  grade,
  gender,
  include = everything(),
  cont_var = NULL
)
}
\arguments{
\item{data}{a \code{tibble} or \code{data.frame}}

\item{path}{a file path to save the excel file}

\item{grade}{specify the grade (year) variable. It must be a factor variable contains '1年', '2年', '3年'.}

\item{gender}{specify the gender variable. It must be a factor variable contains '男', '女'.}

\item{include}{variables to include in the table. Specify the variables in tidy-select style like \code{dplyr::select()}. Default is \code{everything()}, e.g. all variables.}

\item{cont_var}{continuous variables to compute mean instead of frequency. Specify the variables in tidy-select style like \code{dplyr::select()}. Default is \code{NULL}.}
}
\description{
調査実習用の単純集計表を作成する関数です。
}
\examples{
library(dplyr)
library(haven)
data <-
  tibble(
    ID = 1:100,
    grade =
      sample(1:3, 100, replace = TRUE, prob = c(0.3, 0.3, 0.4)) |>
      labelled_spss(labels = c('1年' = 1, '2年' = 2, '3年' = 3), label = '学年'),
    gender =
      sample(c(1, 2, 9), 100, replace = TRUE, prob = c(0.45, 0.45, 0.1)) |>
      labelled_spss(labels = c('男' = 1, '女' = 2, '無回答' = 9), label = '性別'),
    Q1 = rnorm(100, mean = 0, sd = 1),
    Q1_na = sample(c('無回答', '非該当', '有効'), 100, replace = TRUE, prob = c(0.1, 0.1, 0.8)),
    Q2 =
      sample(c(1:4, 8, 9), 100, replace = TRUE, prob = c(0.2, 0.3, 0.3, 0.1, 0.1, 0.1)) |>
      labelled_spss(
        labels = c('あてはまる' = 1, 'まああてはまる' = 2, 'あまりあてはまらない' = 3, 'まったくあてはまらない' = 4, '無回答' = 9, '非該当' = 8),
        label = 'あなたは猫好きである'
      ),
  ) |>
  mutate(
    Q1 = case_when(
      Q1_na == '無回答' ~ 9,
      Q1_na == '非該当' ~ 8,
      .default = Q1
    ) |>
      labelled_spss(labels = c('非該当' = 8, '無回答' = 9), label = 'あなたの猫度を答えてください')
  ) |>
  select(!Q1_na)

data

count(data, grade)
count(data, gender)
count(data, Q1, sort = TRUE)
count(data, Q2)

\dontrun{
jisshu_simple_tabulation(
  data |> as_factor(),
  'neko.xlsx',
  grade = grade,
  gender = gender,
  include = !c(ID),
  cont_var = c(Q1)
)
}

}
