% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisshu_reg.R, R/jisshu_reg.glm.R,
%   R/jisshu_reg.lm.R, R/jisshu_reg.lmer.R
\name{jisshu_reg}
\alias{jisshu_reg}
\alias{jisshu_reg.glm}
\alias{jisshu_reg.lm}
\alias{jisshu_reg.lmer}
\title{Regression Table for Chosa Jisshu}
\usage{
jisshu_reg(object)

\method{jisshu_reg}{glm}(object)

\method{jisshu_reg}{lm}(object)

\method{jisshu_reg}{lmer}(object)
}
\arguments{
\item{object}{an object for which a summary is desired. For example, a model object, such as those returned by \code{lm} and \code{glm}.}
}
\description{
調査実習用の回帰分析結果を作成する関数です。
}
\details{
\code{lm}や\code{glm}の結果から、執筆要項に沿ったエクセルの表が作成できます。

エクセルへの保存はexampleを参照してください。
}
\examples{
# Create a sample data
data <- tibble::tibble(
  x1 = rnorm(100, mean = 0, sd = 1),
  x2 = rnorm(100, mean = 0, sd = 1),
  y = 0.2 + 0.3*x1 + 0.5*x2 + rnorm(100, mean = 0, sd = 0.1),
  y_bin = rbinom(100, 1, plogis(y))
)

# Create a regression table
# Linear regression
model_lm <- lm(y ~ x1 + x2, data = data)
result_lm <- jisshu_reg(model_lm)

# Logistic regression
model_glm <- glm(y_bin ~ x1 + x2, data = data, family = binomial(link = 'logit'))
result_glm <- jisshu_reg(model_glm)

# Save the regression table as an excel file
\dontrun{
result_lm$save('hoge_lm.xlsx')
result_glm$save('hoge_glm.xlsx')
}

}
