% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jisshu_reg.R, R/jisshu_reg.glm.R,
%   R/jisshu_reg.lm.R, R/jisshu_reg.multinom.R
\name{jisshu_reg}
\alias{jisshu_reg}
\alias{jisshu_reg.glm}
\alias{jisshu_reg.lm}
\alias{jisshu_reg.multinom}
\title{Regression Table for Chosa Jisshu}
\usage{
jisshu_reg(object)

\method{jisshu_reg}{glm}(object)

\method{jisshu_reg}{lm}(object)

\method{jisshu_reg}{multinom}(object)
}
\arguments{
\item{object}{an object for which a summary is desired. For example, a model object, such as those returned by \code{lm}, \code{glm}, and \code{multinom}.}
}
\description{
調査実習用の回帰分析結果を作成する関数です。
}
\details{
\code{lm}や\code{glm}、\code{nnet::multinom}の結果から、執筆要項に沿ったエクセルの表が作成できます。

エクセルへの保存はexampleを参照してください。
}
\examples{
data <- tibble::tibble(
  x1 = rnorm(100, mean = 0, sd = 1),
  x2 = rnorm(100, mean = 0, sd = 1),
  y = 0.2 + 0.3*x1 + 0.5*x2 + rnorm(100, mean = 0, sd = 0.1),
  y_bin = rbinom(100, 1, plogis(y)),
  y_multinom = cut(
    y,
    breaks = quantile(y, probs = c(0, 0.25, 0.75, 1)),
    labels = c('Q1', 'Q2', 'Q3'),
    include.lowest = TRUE
  )
)

data

# Create a regression table
# Linear regression
model_lm <- lm(y ~ x1 + x2, data = data)
result_lm <- jisshu_reg(model_lm)

# Logistic regression
model_glm <- glm(y_bin ~ x1 + x2, data = data, family = binomial(link = 'logit'))
result_glm <- jisshu_reg(model_glm)

# Multinomial logistic regression
model_multinom <- nnet::multinom(y_multinom ~ x1 + x2, data = data, trace = FALSE)
result_multinom <- jisshu_reg(model_multinom)

# Glance the regression table
\dontrun{
  result_lm$open()
  result_glm$open()
  result_multinom$open()
}

# Save the regression table as an excel file
\dontrun{
  result_lm$save('hoge_lm.xlsx')
  result_glm$save('hoge_glm.xlsx')
  result_multinom$save('hoge_multinom.xlsx')
}

}
